requirements.txt
PIL
numpy
matplotlib
tqdm
tensorflow-gpu
keras
wanbd
gradio